requirements.txt
absl-py
alembic
appdirs
attrs
cachetools
certifi
cffi
chardet
click
cloudpickle
cryptography
cycler
databricks-cli
docker
entrypoints
Flask
future
gitdb
GitPython
google-auth
google-auth-oauthlib
gorilla
grpcio
idna
isodate
itsdangerous
Jinja2
kiwisolver
Mako
matplotlib
msrest
numpy
oauthlib
pandas
pathspec
Pillow
pio
prometheus-client
prometheus-flask-exporter
protobuf
pyasn1
pyasn1-modules
pycparser
pyparsing
python-dateutil
python-editor
pytorch-lightning
pytz
PyYAML
querystring-parser
regex
requests
requests-oauthlib
rsa
scipy
six
smmap
SQLAlchemy
sqlparse
tabulate
tensorboard
tensorboard-plugin-wit
toml
torch
torchvision
tqdm
typed-ast
typing-extensions
urllib3
waitress
websocket-client
Werkzeug
